keyence_laser/requirements.txt
pyserial==3.5
numpy==1.22.3
matplotlib==3.5.1
pycairo>=1.11.0